+++ projects/ngneat/cashew/src/lib/filterParams.ts
@@ -1,11 +1,11 @@
 import { HttpCacheRequest } from './types';
 
-const filterKeys = ['ttl$', 'cache$', 'key$', 'bucket$'];
+const filterKeys = ['ttl$', 'cache$', 'key$', 'bucket$', 'parameterCodec$'];
 
 export function filterParams(request: HttpCacheRequest) {
   return request.params.keys().reduce((acc, key) => {
     if (filterKeys.includes(key) === false) {
-      acc[key] = request.params.get(key);
+      acc[key] = request.params.getAll(key);
     }
 
     return acc;