+++ projects/ngneat/cashew/src/lib/httpCacheConfig.ts
@@ -1,3 +1,4 @@
+import { HttpParameterCodec } from '@angular/common/http';
 import { InjectionToken } from '@angular/core';
 import { CacheBucket } from './cacheBucket';
 
@@ -5,29 +6,26 @@ export interface HttpCacheConfig {
   strategy: 'implicit' | 'explicit';
   ttl: number;
   responseSerializer?: (value: any) => any;
+  localStorageKey?: string;
+  parameterCodec?: HttpParameterCodec;
 }
 
 export const defaultConfig: HttpCacheConfig = {
   strategy: 'explicit',
-  ttl: 3600000 // One hour
+  ttl: 3600000, // One hour
+  localStorageKey: 'httpCache'
 };
-
-export function mergeConfig(config: Partial<HttpCacheConfig>) {
-  return {
-    ...defaultConfig,
-    ...config
-  };
-}
 
 type Params = {
   cache$?: boolean;
   ttl$?: number;
   key$?: string;
   bucket$?: CacheBucket;
+  parameterCodec$?: HttpParameterCodec;
   [key: string]: any;
 };
 
-export function withCache(params: Params = {}): any {
+export function withCache(params: Params = {}): { params: Params } {
   return {
     params: {
       cache$: true,