+++ projects/ngneat/cashew/src/lib/httpCacheInterceptor.module.ts
@@ -3,7 +3,7 @@ import { ModuleWithProviders, NgModule } from '@angular/core';
 import { HttpCacheInterceptor } from './httpCacheInterceptor';
 import { DefaultKeySerializer, KeySerializer } from './keySerializer';
 import { DefaultHttpCacheStorage, HttpCacheStorage } from './httpCacheStorage';
-import { HTTP_CACHE_CONFIG, HttpCacheConfig, mergeConfig } from './httpCacheConfig';
+import { defaultConfig, HTTP_CACHE_CONFIG, HttpCacheConfig } from './httpCacheConfig';
 import { HttpCacheManager } from './httpCacheManager.service';
 import { DefaultTTLManager, TTLManager } from './ttlManager';
 import { DefaultHttpCacheGuard, HttpCacheGuard } from './httpCacheGuard';
@@ -14,7 +14,7 @@ export class HttpCacheInterceptorModule {
   static forRoot(config: Partial<HttpCacheConfig> = {}): ModuleWithProviders {
     return {
       providers: [
-        { provide: HTTP_CACHE_CONFIG, useValue: mergeConfig(config) },
+        { provide: HTTP_CACHE_CONFIG, useValue: { ...defaultConfig, ...config } },
         { provide: KeySerializer, useClass: DefaultKeySerializer },
         { provide: HttpCacheStorage, useClass: DefaultHttpCacheStorage },
         { provide: TTLManager, useClass: DefaultTTLManager },