+++ projects/ngneat/cashew/src/lib/httpCacheInterceptor.ts
@@ -1,7 +1,8 @@
-import { Injectable } from '@angular/core';
+import { Inject, Injectable } from '@angular/core';
 import { HttpEvent, HttpHandler, HttpInterceptor, HttpRequest, HttpResponse } from '@angular/common/http';
 import { Observable, of } from 'rxjs';
-import { share, tap } from 'rxjs/operators';
+import { finalize, share, tap } from 'rxjs/operators';
+import { HTTP_CACHE_CONFIG, HttpCacheConfig } from './httpCacheConfig';
 
 import { HttpCacheManager } from './httpCacheManager.service';
 import { cloneWithoutParams } from './cloneWithoutParams';
@@ -10,7 +11,11 @@ import { CacheBucket } from './cacheBucket';
 
 @Injectable()
 export class HttpCacheInterceptor implements HttpInterceptor {
-  constructor(private httpCacheManager: HttpCacheManager, private keySerializer: KeySerializer) {}
+  constructor(
+    private httpCacheManager: HttpCacheManager,
+    private keySerializer: KeySerializer,
+    @Inject(HTTP_CACHE_CONFIG) private config: HttpCacheConfig
+  ) {}
 
   intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
     const canActivate = this.httpCacheManager._canActivate(request);
@@ -19,7 +24,11 @@ export class HttpCacheInterceptor implements HttpInterceptor {
     const customKey = request.params.get('key$');
     const bucket: any = request.params.get('bucket$');
 
-    const clone = cloneWithoutParams(request, customKey);
+    const localParameterCodec: any = request.params.get('parameterCodec$');
+    const globalParameterCodec = this.config.parameterCodec;
+    const parameterCodec = localParameterCodec || globalParameterCodec;
+
+    const clone = cloneWithoutParams(request, customKey, parameterCodec);
     const key = this.keySerializer.serialize(clone);
     const queue = this.httpCacheManager._getQueue();
 
@@ -38,8 +47,10 @@ export class HttpCacheInterceptor implements HttpInterceptor {
           if (event instanceof HttpResponse) {
             const cache = this.httpCacheManager._resolveResponse(event);
             this.httpCacheManager._set(key, cache, +ttl);
-            queue.delete(key);
           }
+        }),
+        finalize(() => {
+          queue.delete(key);
         }),
         share()
       );