+++ CHANGELOG.md
@@ -2,6 +2,76 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.3.0](https://github.com/ngneat/cashew/compare/v1.2.0...v1.3.0) (2020-06-18)
+
+
+### Features
+
+* add local parameter codec override ([80148d6](https://github.com/ngneat/cashew/commit/80148d63d9cf79536e21489365208e7dcddeac74))
+* add tests for request parameter codec ([dbc17a6](https://github.com/ngneat/cashew/commit/dbc17a68088815d99f897a75a346d06d45ae3c76))
+* update Params type to be more specific ([c44daf1](https://github.com/ngneat/cashew/commit/c44daf13bad7b8d52e3f8a57d973c5fadb686513))
+
+## [1.2.0](https://github.com/ngneat/cashew/compare/v1.1.4...v1.2.0) (2020-05-28)
+
+
+### Features
+
+* add http parameter codec override ([a1f1a8d](https://github.com/ngneat/cashew/commit/a1f1a8d41b3cd03095aad1a2843428a3dd157d97))
+
+
+### Bug Fixes
+
+* 🐛 support array http parameters ([5ef58d0](https://github.com/ngneat/cashew/commit/5ef58d064f22171d0fb068d0830f0816424d0f3e))
+
+### [1.1.5](https://github.com/ngneat/cashew/compare/v1.1.4...v1.1.5) (2020-05-04)
+
+
+### Bug Fixes
+
+* 🐛 support array http parameters ([5ef58d0](https://github.com/ngneat/cashew/commit/5ef58d064f22171d0fb068d0830f0816424d0f3e))
+
+### [1.1.4](https://github.com/ngneat/cashew/compare/v1.1.3...v1.1.4) (2020-04-29)
+
+
+### Bug Fixes
+
+* remove requests that error from queue ([17b6b0b](https://github.com/ngneat/cashew/commit/17b6b0bd1c0da5d4c5cd2cc2a6b7a1596cda10a0))
+
+### [1.1.3](https://github.com/ngneat/cashew/compare/v1.1.2...v1.1.3) (2020-02-12)
+
+
+### Bug Fixes
+
+* 🐛 aot build ([0ed2eba](https://github.com/ngneat/cashew/commit/0ed2ebac2f78ed15b6771e50fc4db6d54170c5e2))
+
+### [1.1.2](https://github.com/ngneat/cashew/compare/v1.1.1...v1.1.2) (2020-02-12)
+
+
+### Bug Fixes
+
+* 🐛 fix aot build ([9425552](https://github.com/ngneat/cashew/commit/9425552ed57e0ceecc551309e9b6d2f9719886f6))
+
+### [1.1.1](https://github.com/ngneat/cashew/compare/v1.1.0...v1.1.1) (2020-02-12)
+
+
+### Bug Fixes
+
+* 🐛 add type to wrapCache ([ff25b4f](https://github.com/ngneat/cashew/commit/ff25b4fce7ed2d4e7780070cfd9233b82e2b1f00))
+
+## [1.1.0](https://github.com/ngneat/cashew/compare/v1.0.0...v1.1.0) (2020-02-11)
+
+
+### Features
+
+* 🎸 add http cache storage ([07bcb4b](https://github.com/ngneat/cashew/commit/07bcb4ba2f56b2b22877a2cb915f9ca8c030e19e))
+* 🎸 Local Storage Option ([0c5f29a](https://github.com/ngneat/cashew/commit/0c5f29a56d189af9b8aca1641e595a847ecbece5))
+
+
+### Tests
+
+* 💍 Added local storage tests ([84ff419](https://github.com/ngneat/cashew/commit/84ff419277b3b282aa055c2f997ca634933b6513))
+* 💍 Memory Use in Local Storage ([1df3852](https://github.com/ngneat/cashew/commit/1df3852803b1cf05c005cc46cf7f8b50c0d19351))
+
 ## 1.0.0 (2020-01-30)